+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,27 @@
 
 ### Patch Changes
 
+- Updated dependencies [9ba4324]
+  - ai@6.0.0-beta.137
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [3d83f38]
+  - ai@6.0.0-beta.136
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [afe7093]
+  - ai@6.0.0-beta.135
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [686103c]
   - ai@6.0.0-beta.134